requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
Pillow==9.4.0
pypng==0.20220715.0
PyQRCode==1.2.1
pytz==2022.7.1
sqlparse==0.4.3
